requirements.txt
asgiref==3.8.1
attrs==25.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
Django==5.2b1
django-environ==0.12.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docutils==0.21.2
drf-spectacular==0.28.0
fonttools==4.54.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
inflection==0.5.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.4.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
matplot==0.1.9
matplotlib==3.9.2
mdurl==0.1.2
more-itertools==10.5.0
mysqlclient==2.2.7
nh3==0.2.18
numpy==2.1.1
packaging==24.1
pandas==2.2.3
pillow==10.4.0
pkginfo==1.10.0
Pygments==2.18.0
PyJWT==2.10.1
pyloco==0.0.139
PyMySQL==1.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
readme_renderer==44.0
referencing==0.36.2
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.8.1
rpds-py==0.23.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
SimpleWebSocketServer==0.1.2
six==1.16.0
sqlparse==0.5.3
threadpoolctl==3.5.0
twine==5.1.1
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
ushlex==0.99.1
websocket-client==1.8.0
zipp==3.20.2
